move server to root
